Change ISBN to BigIntegerField

+++ post_textbook/models.py
@@ -26,7 +26,7 @@ class Textbook(models.Model):
     publisher = models.CharField(max_length=200)
     edition = models.IntegerField()
     year = models.IntegerField()
-    ISBN = models.IntegerField()
+    ISBN = models.BigIntegerField()
     cover = models.ImageField(upload_to=None)
     
     def __str__(self):
